apps/backend/pyproject.toml
[tool.poetry]
name = "aibat-backend"
version = "0.1.0"
description = ""
authors = ["Anurag Maravi <anuragmaravi12@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
firebase-admin = "^6.9.0"
python-dotenv = "^1.1.1"
pandas = "^2.3.0"
requests = "^2.31.0"
vertexai = "^1.38.0"
google-cloud-aiplatform = "^1.101.0"
